requirements.txt
streamlit
numpy
pydicom
gdown
tensorflow
Pillow
opencv-python-headless
scikit-image
scipy